requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
python-slugify==8.0.1
Markdown==3.4.4
bleach==6.0.0
pymysql==1.1.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.0.0
email-validator==2.0.0
